Moved common functions in common.py
