+++ zxtools/__init__.py
@@ -7,35 +7,7 @@
 # Licensed under the BSD 3-Clause license.
 # See LICENSE file in the project root for full license information.
 #
-"""Global constants and templates are here"""
-
-import sys
-import logging
+"""Global constants"""
 
 __version__ = '1.0.23'
 CHUNK_SIZE = 512 * 1024  # 512 KBytes
-
-
-def safe_parse_args(parser, args):
-    """Safely parse arguments"""
-    try:
-        options = parser.parse_args(args)
-        if len(args) == 0:
-            raise ValueError
-    except ValueError:
-        parser.print_help()
-        sys.exit(0)
-
-    return options
-
-
-def default_main(parser):
-    """ Default entry point implementation """
-    args = safe_parse_args(parser, sys.argv[1:])
-    if args.verbose:
-        logging.basicConfig(level=logging.DEBUG)
-
-    if hasattr(args, 'func'):
-        args.func(args)
-
-    return args
